+++ weldx/asdf/util.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import importlib.metadata
 from collections.abc import Callable, Hashable, Mapping, MutableMapping, Set
 from contextlib import contextmanager
 from io import BytesIO, TextIOBase
@@ -153,7 +154,7 @@ def read_buffer_context(
         Buffer containing ASDF file contents
     open_kwargs
         Additional keywords to pass to `asdf.AsdfFile.open`
-        Extensions are always set, ``copy_arrays=True`` is set by default.
+        Extensions are always set, ``memmap=False`` is set by default.
 
     Returns
     -------
@@ -162,7 +163,13 @@ def read_buffer_context(
 
     """
     if open_kwargs is None:
-        open_kwargs = {"copy_arrays": True, "lazy_load": False}
+        open_kwargs = {"memmap": False, "lazy_load": False}
+
+    if "memmap" in open_kwargs and tuple(
+        importlib.metadata.version("asdf").split(".")
+    ) < ("3", "1", "0"):
+        open_kwargs["copy_arrays"] = not open_kwargs["memmap"]
+        del open_kwargs["memmap"]
 
     buffer.seek(0)
 
@@ -194,7 +201,7 @@ def read_buffer(
         Buffer containing ASDF file contents
     open_kwargs
         Additional keywords to pass to `asdf.AsdfFile.open`
-        Extensions are always set, ``copy_arrays=True`` is set by default.
+        Extensions are always set, ``memmap=False`` is set by default.
 
     Returns
     -------
@@ -224,7 +231,7 @@ def write_read_buffer_context(
         Extensions are always set.
     open_kwargs
         Additional keywords to pass to `asdf.AsdfFile.open`
-        Extensions are always set, ``copy_arrays=True`` is set by default.
+        Extensions are always set, ``memmap=False`` is set by default.
 
     Returns
     -------
@@ -252,7 +259,7 @@ def write_read_buffer(
         Extensions are always set.
     open_kwargs
         Additional keywords to pass to `asdf.AsdfFile.open`
-        Extensions are always set, ``copy_arrays=True`` is set by default.
+        Extensions are always set, ``memmap=False`` is set by default.
 
     Returns
     -------