+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 ### Fixes
 
 - rename (fix typo) argument to `lcs_child_in_parent` in `CoordinateSystemManager.add_cs`  \[{pull}`936`\].
+- replace usages of `pkg_resources` with `importlib.metadata` \[{pull}`941`\].
+- replace usages of `copy_arrays` with `memmap` for `asdf>=3.1.0` \[{pull}`940`\].
 
 ### Dependencies
 
 - pin `weldx-widgets>=0.2.3` for viz \[{pull}`939`\].
+- pin `pint>=0.21` \[{pull}`941`\].
 
 ## 0.6.8 (07.06.2024)